+++ packages/paste-core/components/paragraph/__tests__/paragraph.test.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import renderer from 'react-test-renderer';
+import {render} from 'react-dom';
+import {axe} from 'jest-axe';
 import {Theme} from '@twilio-paste/theme';
 import {Paragraph} from '../src';
 
@@ -7,7 +9,7 @@ describe('Paragraph', () => {
   it('it should render a single paragraph', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>This is a paragraph</Paragraph>
         </Theme.Provider>
       )
@@ -15,10 +17,23 @@ describe('Paragraph', () => {
     expect(tree).toMatchSnapshot();
   });
 
+  it('should have no accessibility violations', async () => {
+    const container = document.createElement('div');
+    document.body.append(container);
+    render(
+      <Theme.Provider theme="console">
+        <Paragraph>This is a paragraph</Paragraph>
+      </Theme.Provider>,
+      container
+    );
+    const results = await axe(document.body);
+    expect(results).toHaveNoViolations();
+  });
+
   it('it should render a single paragraph with italic text', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>
             <i>This is a paragraph with italic text</i>
           </Paragraph>
@@ -31,7 +46,7 @@ describe('Paragraph', () => {
   it('it should render a single paragraph with bold text', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>
             <strong>This is a paragraph with bold text</strong>
           </Paragraph>