+++ packages/paste-style-props/src/typography.ts
@@ -1,0 +1,13 @@
+export const TYPOGRAPHY_PROPS = [
+  'fontFamily',
+  'fontSize',
+  'fontStyle',
+  'fontWeight',
+  'letterSpacing',
+  'lineHeight',
+  'textAlign',
+  'textColor',
+  'textDecoration',
+  'textOverflow',
+  'whiteSpace',
+];