+++ packages/paste-website/src/components/shortcodes/component-header/index.tsx
@@ -42,6 +42,7 @@ interface ComponentHeaderProps {
   categoryRoute: typeof SidebarCategoryRoutes[keyof typeof SidebarCategoryRoutes];
   githubUrl: string;
   storybookUrl: string;
+  abstractUrl: string;
   data?: [
     {
       node: {
@@ -80,7 +81,14 @@ const PackageInstallSnippet: React.FC<{}> = ({children}) => {
   );
 };
 
-const ComponentHeader: React.FC<ComponentHeaderProps> = ({name, categoryRoute, githubUrl, storybookUrl, data}) => {
+const ComponentHeader: React.FC<ComponentHeaderProps> = ({
+  name,
+  categoryRoute,
+  githubUrl,
+  storybookUrl,
+  abstractUrl,
+  data,
+}) => {
   if (data == null || data[0] == null || data[0].node == null) {
     return <ComponentHeaderBasic categoryRoute={categoryRoute} name={name} />;
   }
@@ -107,7 +115,10 @@ const ComponentHeader: React.FC<ComponentHeaderProps> = ({name, categoryRoute, g
             <Box display="inline" marginRight="space30">
               <Anchor href={githubUrl}>Github</Anchor>
             </Box>
-            {storybookUrl != null ? <Anchor href={storybookUrl}>Storybook</Anchor> : null}
+            <Box display="inline" marginRight="space30">
+              {storybookUrl != null ? <Anchor href={storybookUrl}>Storybook</Anchor> : null}
+            </Box>
+            {abstractUrl != null ? <Anchor href={abstractUrl}>Abstract</Anchor> : null}
           </PackageValue>
         </Box>
         <Box marginBottom="space20">