+++ packages/paste-core/components/heading/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@1.0.6...@twilio-paste/heading@1.0.7) (2019-12-20)
+
+**Note:** Version bump only for package @twilio-paste/heading
+
+
+
+
+
+## [1.0.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@1.0.5...@twilio-paste/heading@1.0.6) (2019-12-19)
+
+**Note:** Version bump only for package @twilio-paste/heading
+
+
+
+
+
 ## [1.0.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@1.0.4...@twilio-paste/heading@1.0.5) (2019-12-05)
 
 **Note:** Version bump only for package @twilio-paste/heading