+++ packages/paste-core/utilities/absolute/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/absolute@2.0.5...@twilio-paste/absolute@2.0.6) (2019-12-20)
+
+**Note:** Version bump only for package @twilio-paste/absolute
+
+
+
+
+
+## [2.0.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/absolute@2.0.4...@twilio-paste/absolute@2.0.5) (2019-12-19)
+
+**Note:** Version bump only for package @twilio-paste/absolute
+
+
+
+
+
 ## [2.0.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/absolute@2.0.3...@twilio-paste/absolute@2.0.4) (2019-12-05)
 
 **Note:** Version bump only for package @twilio-paste/absolute